chap06 정리

+++ java-cc/chap02_datatypeandprimarydatatype-master/sec02_primarydatatype/EX04_TypeCasting_1/TypeCasting_1.java
@@ -1,6 +1,6 @@
 package sec02_primarydatatype.EX04_TypeCasting_1;
 
-/*두 가지 타입 변환 방법*/
+/* 두 가지 타입 변환 방법 */
 
 public class TypeCasting_1 {
 
@@ -10,9 +10,9 @@ public class TypeCasting_1 {
 
 		// @1-1 캐스팅 방법 1 (데이터타입)
 		int value1 = (int) 5.3;
-		long value2 = (long) 10;
+		long value2 = 10;
 		float value3 = (float) 5.8;
-		double value4 = (double) 16;
+		double value4 = 16;
 
 		System.out.println(value1); // 5
 		System.out.println(value2); // 10
